app/requirements.txt
flask
mysql-connector-python
pika==1.3.2
psycopg2-binary==2.9.7
pymongo==4.5.0